Add support of electron ipc to communicator. Add chrome debugging support in vscode.
