+++ src/Sfx-Standalone/modules/ipc/index.ts
@@ -9,7 +9,7 @@ import { ILog } from "sfx.logging";
 import { ChildProcess } from "child_process";
 import { Socket } from "net";
 
-import { NodeCommunicator } from "./communicator.node";
+import { Communicator } from "./communicator";
 import { electron } from "../../utilities/electron-adapter";
 
 export function getModuleMetadata(): IModuleInfo {
@@ -19,10 +19,10 @@ export function getModuleMetadata(): IModuleInfo {
         loadingMode: "Always",
         components: [
             {
-                name: "ipc.communicator-node",
+                name: "ipc.communicator",
                 version: electron.app.getVersion(),
                 deps: ["logging"],
-                descriptor: (log: ILog, channel: NodeJS.Process | ChildProcess | Socket, id?: string) => new NodeCommunicator(channel, id)
+                descriptor: (log: ILog, channel: NodeJS.Process | ChildProcess | Socket, id?: string) => new Communicator(channel, id)
             }
         ]
     };
