+++ src/Sfx-Standalone/modules/prompt/preload.ts
@@ -6,7 +6,7 @@
 import { PromptContext } from "./prompt-context";
 import { electron } from "../../utilities/electron-adapter";
 import * as mmutils from "../../module-manager/utils";
-import { NodeCommunicator } from "../ipc/communicator.node";
+import { Communicator } from "../ipc/communicator";
 import * as utils from "../../utilities/utils";
 import { ChannelNameFormat, EventNames } from "./constants";
 
@@ -15,7 +15,7 @@ process.once("loaded", async () => {
     const constructorOptions =
         electron.ipcRenderer.sendSync(utils.format(ChannelNameFormat, promptWindow.id, EventNames.RequestModuleManagerConstructorOptions));
 
-    global["sfxModuleManager"] = await mmutils.createModuleManagerAsync(constructorOptions, new NodeCommunicator(process));
+    global["sfxModuleManager"] = await mmutils.createModuleManagerAsync(constructorOptions, new Communicator(process));
 
     sfxModuleManager.registerComponents([
         {
